Add CLI configuration flags.
Added a simple argparser object to confiure
the geth exporter using the commandline.

It overrides default values and settings in a
yml file (if present), but only does so for
explicitly stated config options. If the
option is not passed, no overriding takes
place and the usual yml>defaults order
takes place.